feat(rust) Base implementation of the rust library (#22)
Introduces the Rust implementation of the usage-accountant library. The python version is in this repo in the py subdirectory.

Context on what the library does:
This is part of the project to break down our cloud usage by feature. Since a lot of our infra is sahred we need to instrument the application to provide usage data broken down by product feature in case of shared resources like Relay, Symbolicator, the Snuba API, the generic metrics platform etc.

The data flows from the application to Kafka to bigquery to be used in analysis.
This library standardizes the way we produce usage data on Kafka to ensure the schema is consistent and that we do not overwhelm the consumer (pre-aggregation).

More details on the usage are in the Crate documentation.